fetch google places in splash
