+++ app/components/screens/auth/splash.js
@@ -7,11 +7,19 @@ import {
 import baseStyles from '../styles/styles'
 
 export default class SplashScreen extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      position: 'unknown',
+    };
+    this.fetchGoogle = this.fetchGoogle.bind(this);
+  }
+
   componentWillMount() {
     navigator.geolocation.getCurrentPosition(
       (position) => {
         var initialPosition = position;
-        this.setState({initialPosition});
+        this.setState(() => {position: initialPosition});
         this.props.receiveLocation(initialPosition);
       },
       (error) => alert(JSON.stringify(error)),
@@ -19,44 +27,54 @@ export default class SplashScreen extends Component {
     );
     this.watchID = navigator.geolocation.watchPosition((position) => {
       var lastPosition = position;
-      this.setState({lastPosition});
       this.props.receiveLocation(lastPosition);
+      this.fetchGoogle(lastPosition.coords.latitude, lastPosition.coords.longitude);
     });
     this.getToken();
   }
 
-    async getToken() {
-      try {
-        let accessToken = await this.props.getItem('token');
-        if(!accessToken) {
-          console.log("Token not set");
-          this.props.navigation.navigate('LogIn');
-        } else {
-          this.verifyToken(accessToken)
-        }
-      } catch(error) {
-        console.log("Something went wrong");
+  async getToken() {
+    try {
+      let accessToken = await this.props.getItem('token');
+      if(!accessToken) {
+        console.log("Token not set");
+        this.props.navigation.navigate('LogIn');
+      } else {
+        this.verifyToken(accessToken)
       }
+    } catch(error) {
+      console.log("Something went wrong");
     }
+  }
 
-    async verifyToken(token) {
-      let accessToken = token
-      try {
-        let response = await fetch('https://later-chat.herokuapp.com/api/verify?session%5Baccess_token%5D='+accessToken);
-        let res = await response.text();
-        if (response.status >= 200 && response.status < 300) {
-          //Verified token means user is logged in so we redirect him to home.
-          this.props.navigation.navigate('Tabs');
-        } else {
-            //Handle error
-            this.props.navigation.navigate('LogIn');
-            let error = res;
-            throw error;
-        }
-      } catch(error) {
-          console.log("error response: " + error);
+  async verifyToken(token) {
+    let accessToken = token
+    try {
+      let response = await fetch('https://later-chat.herokuapp.com/api/verify?session%5Baccess_token%5D='+accessToken);
+      let res = await response.text();
+      if (response.status >= 200 && response.status < 300) {
+        //Verified token means user is logged in so we redirect him to home.
+        this.props.navigation.navigate('Tabs');
+      } else {
+        //Handle error
+        this.props.navigation.navigate('LogIn');
+        let error = res;
+        throw error;
       }
+    } catch(error) {
+      console.log("error response: " + error);
     }
+  }
+
+  async fetchGoogle(lat, lng) {
+    let response = await fetch('https://maps.googleapis.com/maps/api/place/nearbysearch/json?location='+lat+','+lng+'&radius=500&key=AIzaSyDBH-I807okFiwNi3VqRYFuHpdOYH4DXX4');
+    let res = await response.json();
+    let places_nearby = [];
+    for (var i = 1; i < 5; i++) {
+      places_nearby.push(res.results[i].name);
+    }
+    this.props.receiveGooglePlaces({places_nearby: places_nearby});
+  }
 
   render () {
